Upload of xAPI Profile Primer FIRST DRAFT
We discussed the thought of creating an xAPI Profile Primer during our team meeting on 10-21-21. This first draft is a starting point for those looking to understand and use the xAPI Profile Server to locate, identify and use recommended profiles as a starting point for their projects.

It was published using the 'xAPI Profile Human-readable Template.docx' file provided by Jason Haag as a template for Human Readable xAPI Profile Documentation.